separated enums and models
