+++ packages/nativescript-toast/enums/color-map.ts
@@ -1,0 +1,6 @@
+export enum ToastColorMap {
+  SuccessText = '#ffffff',
+  SuccessBackground = '#4e576b',
+  ErrorText = '#e12a37',
+  ErrorBackground = '#fef7f7',
+}
