+++ packages/nativescript-toast/enums/toast-duration.ts
@@ -1,0 +1,4 @@
+export enum ToastDuration {
+  'Short' = 'short',
+  'Long' = 'long'
+}
