+++ packages/nativescript-toast/enums/toast-variant.ts
@@ -1,0 +1,4 @@
+export enum ToastVariant {
+  'Success' = 'success',
+  'Error' = 'error',
+}
