+++ packages/nativescript-toast/index.d.ts
@@ -1,67 +1,10 @@
-import { Color } from '@nativescript/core';
-import { ToastDuration, ToastVariant, ToastVariantParams } from './toast.common';
+import { ToastOptions } from './models/toast-options';
 
 export declare class Toasty {
-  constructor(opts: ToastyOptions);
-
-  duration: ToastDuration;
-  variant: ToastVariant;
-  customVariantParams: ToastVariantParams;
+  constructor(opts: ToastOptions);
 
   /**
    * Shows the toast.
    */
   show(): void;
-}
-
-export interface ToastyOptions {
-  /**
-   * Message text of the Toast.
-   */
-  text: string;
-
-  /**
-   * Duration to show the Toast.
-   */
-  duration?: ToastDuration;
-
-  /**
-   * Change Toast Variant - default SUCCESS
-   */
-  variant?: ToastVariant;
-
-  /**
-   * Set specific background and text color
-   */
-  customVariantParams?: ToastVariantParams;
-
-  /**
-   * Android specific configuration options.
-   */
-  android?: any;
-
-  /**
-   * iOS Specific configuration options.
-   */
-  ios?: {
-    /**
-     * The number of lines to allow for the toast message.
-     */
-    messageNumberOfLines?: number;
-
-    /**
-     * The corner radius of the Toast.
-     */
-    cornerRadius?: number;
-
-    /**
-     * True to display a shadow for the Toast.
-     */
-    displayShadow?: boolean;
-
-    /**
-     * The color of the shadow. Only visible if `displayShadow` is true.
-     */
-    shadowColor?: Color | string;
-  };
 }
