+++ packages/nativescript-toast/models/variant-params.ts
@@ -1,0 +1,4 @@
+export interface ToastVariantParams {
+  backgroundColor: string;
+  textColor: string;
+}
